Move files from transfer to store
