+++ peernode/src/peer-node/server.rs
@@ -2,7 +2,6 @@ pub mod consumer;
 pub mod peer;
 pub mod producer;
 pub mod store;
-pub mod transfer;
 
 pub mod jobs;
