--- peernode/src/peer-node/transfer/mod.rs
@@ -1,1 +1,0 @@
-pub mod files;
